requirements.txt
selenium
apscheduler
webdriver-manager
whatsapp-web.js
qrcode-terminal